Move YiiDevToolApplication to App
